+++ mail/mail-tui/src/widgets/utils.rs
@@ -1,5 +1,7 @@
 use chrono::DateTime;
-use proton_mail_common::datatypes::{MessageAddress, MessageAddresses};
+use proton_mail_common::datatypes::{
+    MessageRecipient, MessageRecipients, MessageSender, MessageSenders,
+};
 
 pub fn date_from_timestamp(timestamp: u64) -> String {
     let timestamp_i64 = i64::try_from(timestamp).unwrap_or(0);
@@ -8,7 +10,7 @@ pub fn date_from_timestamp(timestamp: u64) -> String {
     let date_str = date.format("%d/%m/%Y %H:%M");
     date_str.to_string()
 }
-pub fn sender_name(sender: &MessageAddress) -> &str {
+pub fn sender_name(sender: &MessageSender) -> &str {
     if sender.name.is_empty() {
         sender.address.as_str()
     } else {
@@ -16,7 +18,7 @@ pub fn sender_name(sender: &MessageAddress) -> &str {
     }
 }
 
-pub fn format_sender(sender: &MessageAddress) -> String {
+pub fn format_sender(sender: &MessageSender) -> String {
     if sender.name.is_empty() {
         sender.address.clone()
     } else {
@@ -24,11 +26,28 @@ pub fn format_sender(sender: &MessageAddress) -> String {
     }
 }
 
-pub fn format_senders(senders: &MessageAddresses) -> String {
+pub fn format_recipient(sender: &MessageRecipient) -> String {
+    if sender.name.is_empty() {
+        sender.address.clone()
+    } else {
+        format!("{} <{}>", sender.name, sender.name)
+    }
+}
+
+pub fn format_senders(senders: &MessageSenders) -> String {
     senders
         .value
         .iter()
         .map(format_sender)
+        .collect::<Vec<_>>()
+        .join(", ")
+}
+
+pub fn format_recipients(senders: &MessageRecipients) -> String {
+    senders
+        .value
+        .iter()
+        .map(format_recipient)
         .collect::<Vec<_>>()
         .join(", ")
 }